Proper handling of default captcha on extension enable/disable/purge
In order to update an extension, it has to be disabled. If Sortables
Captcha is set as the default captcha, the board will break because it
can't find the service anymore. On disabling the extension phpBB's
default GD captcha will be restored and a remember flag is added to
automatically reinstate Sortables Captcha as the default captcha when
the extension is re-enabled.